+++ uui/src/lib/index.ts
@@ -2,6 +2,7 @@
 import Box from './Box/box.svelte';
 import Chip from './Chip/chip.svelte';
 import Step from './Stepper/Step.svelte';
+import List from './List/List.svelte';
 import Paper from './Paper/Paper.svelte';
 import Overlay from './Paper/Overlay.svelte';
 import Theme from './Theme/theme.svelte';
@@ -11,19 +12,25 @@ import Stepper from './Stepper/Stepper.svelte';
 import ChipIcon from './Chip/chip-icon.svelte';
 import StepIcon from './Stepper/StepIcon.svelte';
 import TextField from './TextField/TextField.svelte';
+import ListItem from './List/ListItem.svelte';
+import ListGroup from './List/ListGroup.svelte';
 import Accordion from './Accordion/accordion.svelte';
 import ButtonIcon from './Button/ButtonIcon.svelte';
 import IconButton from './IconButton/IconButton.svelte';
 import Typography from './Typography/typography.svelte';
 import DialogTitle from './Dialog/DialogTitle.svelte';
-import TextFieldIcon from './TextField/TextFieldIcon.svelte';
+import ListSubHeader from './List/ListSubHeader.svelte';
 import DialogActions from './Dialog/DialogActions.svelte';
+import TextFieldIcon from './TextField/TextFieldIcon.svelte';
 import AccordionGroup from './Accordion/accordion-group.svelte';
+import ListItemCheckbox from './List/ListItemCheckbox.svelte';
+import ListItemAdornment from './List/ListItemAdornment.svelte';
 
 export {
   Box,
   Chip,
   Step,
+  List,
   Paper,
   Overlay,
   Theme,
@@ -32,15 +39,20 @@ export {
   Stepper,
   ChipIcon,
   StepIcon,
+  ListItem,
+  ListGroup,
   TextField,
   Accordion,
   ButtonIcon,
   IconButton,
   Typography,
   DialogTitle,
+  ListSubHeader,
   TextFieldIcon,
   DialogActions,
-  AccordionGroup
+  AccordionGroup,
+  ListItemCheckbox,
+  ListItemAdornment
 };
 
 // Export Utility Functions here
@@ -103,6 +115,16 @@ export type {
 export type { TextFieldProps, TextFieldIconProps } from './TextField/TextField.types';
 export type { AccordionProps, AccordionGroupProps } from './Accordion/accordion.types';
 export type { ThemeColors } from './shared/theme/default-colors/colors.types';
+export type { ComponentAsProp } from './shared/types/ComponentAsProp';
+export type { ListGroupProps } from './List/ListGroup.types';
+export type { ListSubHeaderProps } from './List/ListSubHeader.types';
+export type { ListProps } from './List/List.types';
+export type {
+  ListItemProps,
+  ListItemSlots,
+  ListItemSlotProps,
+  ListItemAdornmentProps
+} from './List/ListItem.types';
 export type {
   Breakpoints,
   BreakpointKey,